Retrieve a list of feeds created by a given actor

Retrieve a list of feeds created by a given actor

Usage

bs_get_feeds(
  actor,
  cursor = NULL,
  limit = NULL,
  user = get_bluesky_user(),
  pass = get_bluesky_pass(),
  auth = bs_auth(user, pass),
  clean = TRUE
)

Arguments

actor

Character, length 1. name of 1 actor, such as 'chriskenny.bsky.social'

cursor

Character, length 1. A cursor property from a prior response. Default: NULL.

limit

Integer. Number of records to request. If over 100, multiple requests are made.

user

Character. User name to log in with. Defaults to get_bluesky_user().

pass

Character. App password to log in with. Defaults to get_bluesky_pass().

auth

Authentication information. Defaults to bs_auth(user, pass).

clean

Logical. Should output be cleaned into a tibble? Default: TRUE.

Value

a tibble::tibble of feeds

Lexicon references

feed/getActorFeeds.json (2023-10-01)

Function introduced

v0.0.1 (2023-10-01)

Examples

if (FALSE) { # has_bluesky_pass() && has_bluesky_user()
bs_get_feeds('chriskenny.bsky.social')
}